重构将代码放在另一个文件
